+++ changelog.md
@@ -1,8 +1,26 @@
+## [0.0.3] - 2022-23-02 - R.S.
+### Ajout
+-Mise en place de l'invitation à un lobby.
+-Mise en place de l'envoie de message vers un lobby.
+-Mise en place de la déconnexion à un lobby.
+-Passage de l'objet lobby en singleton.
+-Évènement lors de l'entrée ou la sortie dans le lobby.
+-Évènement lors de la réception d'un message dans le lobby.
+-Évènement lors de la modification des données d'un lobby.
+-Évènement lors de la réception d'une invitation à un lobby.
+
 ## [0.0.2] - 2022-22-02 - R.S.
 ### Ajout
--Mise en place de la réponse au message.
--Mise en place de la création et connexion à des lobbys.
+-Création d'un début d'interface pour un écran 16:9.
+-Mise en place de la réponse au message d'ami steam.
+-Mise en place de la création d'un lobby.
+-Mise en place de la connexion à un lobby.
+-Mise en place du filtre pour la liste des lobbys.
+-Évènement lors de la réception d'une liste de lobbys.
+-Évènement lors de la création d'un lobby.
+-Évènement lors de la connexion à un lobby.
+-Évènement lors de la réception d'une message d'ami steam.
 
 ## [0.0.1] - 2022-21-02 - R.S.
 ### Ajout
--API Steam pour Unity.
+-Mise en place et test de l'API Steam pour Unity.